# core/docs/swagger.yaml
...
info:
  contact: {}
  title: AI Proxy Swagger API
  version: "1.0"
paths:
...
  /api/log/{group}/export:
    get:
      description: Streams filtered group logs as a CSV table file
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Start timestamp, max span 30 days
        in: query
        name: start_timestamp
        type: integer
      - description: End timestamp, max span 30 days
        in: query
        name: end_timestamp
        type: integer
      - description: Model name
        in: query
        name: model_name
        type: string
      - description: Token ID
        in: query
        name: token_id
        type: integer
      - description: Token name
        in: query
        name: token_name
        type: string
      - description: Sort order for created_at, supports desc or asc
        in: query
        name: order
        type: string
      - description: Request ID
        in: query
        name: request_id
        type: string
      - description: Upstream ID
        in: query
        name: upstream_id
        type: string
      - description: Status code type
        in: query
        name: code_type
        type: string
      - description: Status code
        in: query
        name: code
        type: integer
      - description: Include request and response detail, default false
        in: query
        name: include_detail
        type: boolean
      - description: IP
        in: query
        name: ip
        type: string
      - description: User
        in: query
        name: user
        type: string
      - description: Timezone, default is Local
        in: query
        name: timezone
        type: string
      - description: Maximum exported rows; zero or negative means unlimited
        in: query
        name: max_entries
        type: integer
      - description: Include channel column, default false
        in: query
        name: include_channel
        type: boolean
      - description: Include retry_at column, default false
        in: query
        name: include_retry_at
        type: boolean
      - description: Chunk interval, default 30m, min 10m, max 4h, e.g. 10m, 30m,
          1h
        in: query
        name: chunk_interval
        type: string
      produces:
      - text/csv
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Export group logs
      tags:
      - log
...
  /api/logs/export:
    get:
      description: Streams filtered global logs as a CSV table file
      parameters:
      - description: Start timestamp, max span 30 days
        in: query
        name: start_timestamp
        type: integer
      - description: End timestamp, max span 30 days
        in: query
        name: end_timestamp
        type: integer
      - description: Model name
        in: query
        name: model_name
        type: string
      - description: Channel ID
        in: query
        name: channel
        type: integer
      - description: Sort order for created_at, supports desc or asc
        in: query
        name: order
        type: string
      - description: Request ID
        in: query
        name: request_id
        type: string
      - description: Upstream ID
        in: query
        name: upstream_id
        type: string
      - description: Status code type
        in: query
        name: code_type
        type: string
      - description: Status code
        in: query
        name: code
        type: integer
      - description: Include request and response detail, default false
        in: query
        name: include_detail
        type: boolean
      - description: IP
        in: query
        name: ip
        type: string
      - description: User
        in: query
        name: user
        type: string
      - description: Timezone, default is Local
        in: query
        name: timezone
        type: string
      - description: Maximum exported rows; zero or negative means unlimited
        in: query
        name: max_entries
        type: integer
      - description: Chunk interval, default 30m, min 10m, max 4h, e.g. 10m, 30m,
          1h
        in: query
        name: chunk_interval
        type: string
      produces:
      - text/csv
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Export global logs
      tags:
      - logs
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"